backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI Backend for minimal OpenAI ChatBot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "openai>=1.99.6",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
